GROW-352 banner changes (#2381)
* Rename to AppealBannerCircular, progress towards new design

* Add arcscale and rotate props to KvProgressCircle to generate true c-shape

* Simplify c shaped circle

* Misc styling

* more styling, fix native button clicks

* more styling, fix native button clicks

* styling
